+++ framework/db/Connection.php
@@ -22,9 +22,9 @@ use yii\caching\CacheInterface;
  * of the [PDO PHP extension](https://secure.php.net/manual/en/book.pdo.php).
  *
  * Connection supports database replication and read-write splitting. In particular, a Connection component
- * can be configured with multiple [[masters]] and [[slaves]]. It will do load balancing and failover by choosing
- * appropriate servers. It will also automatically direct read operations to the slaves and write operations to
- * the masters.
+ * can be configured with multiple [[primaries]] and [[replicas]]. It will do load balancing and failover by choosing
+ * appropriate servers. It will also automatically direct read operations to the replicas and write operations to
+ * the primary connections.
  *
  * To establish a DB connection, set [[dsn]], [[username]] and [[password]], and then
  * call [[open()]] to connect to the database server. The current state of the connection can be checked using [[$isActive]].
@@ -114,21 +114,21 @@ use yii\caching\CacheInterface;
  * @property bool $isActive Whether the DB connection is established. This property is read-only.
  * @property string $lastInsertID The row ID of the last row inserted, or the last value retrieved from the
  * sequence object. This property is read-only.
- * @property Connection $master The currently active master connection. `null` is returned if there is no
- * master available. This property is read-only.
- * @property PDO $masterPdo The PDO instance for the currently active master connection. This property is
+ * @property-read Connection|null $primary The currently active primary connection. `null` is returned if no primary
+ * connection is available. This property is read-only.
+ * @property-read PDO $primaryPdo The PDO instance for the currently active primary connection. This property is
  * read-only.
  * @property QueryBuilder $queryBuilder The query builder for the current DB connection. Note that the type of
  * this property differs in getter and setter. See [[getQueryBuilder()]] and [[setQueryBuilder()]] for details.
  * @property Schema $schema The schema information for the database opened by this connection. This property
  * is read-only.
  * @property string $serverVersion Server version as a string. This property is read-only.
- * @property Connection $slave The currently active slave connection. `null` is returned if there is no slave
- * available and `$fallbackToMaster` is false. This property is read-only.
- * @property PDO $slavePdo The PDO instance for the currently active slave connection. `null` is returned if
- * no slave connection is available and `$fallbackToMaster` is false. This property is read-only.
+ * @property-read Connection $replica The currently active replica connection. This property is read-only.
+ * @property-read PDO $replicaPdo The PDO instance for the currently active replica connection. This property
+ * is read-only.
  * @property Transaction|null $transaction The currently active transaction. Null if no active transaction.
  * This property is read-only.
+ * @mixin ConnectionDeprecationsTrait
  *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
@@ -329,79 +329,195 @@ class Connection extends Component
     public $enableSavepoint = true;
     /**
      * @var CacheInterface|string|false the cache object or the ID of the cache application component that is used to store
-     * the health status of the DB servers specified in [[masters]] and [[slaves]].
-     * This is used only when read/write splitting is enabled or [[masters]] is not empty.
+     * the health status of the DB servers specified in [[primaries]] and [[replicas]].
+     * This is used only when read/write splitting is enabled or [[primaries]] is not empty.
      * Set boolean `false` to disabled server status caching.
      * @see openFromPoolSequentially() for details about the failover behavior.
      * @see serverRetryInterval
      */
     public $serverStatusCache = 'cache';
     /**
-     * @var int the retry interval in seconds for dead servers listed in [[masters]] and [[slaves]].
+     * @var int the retry interval in seconds for dead servers listed in [[primaries]] and [[replicas]].
      * This is used together with [[serverStatusCache]].
      */
     public $serverRetryInterval = 600;
     /**
-     * @var bool whether to enable read/write splitting by using [[slaves]] to read data.
-     * Note that if [[slaves]] is empty, read/write splitting will NOT be enabled no matter what value this property takes.
+     * @var bool whether to enable read/write splitting by using [[replicas]] to read data.
+     * Note that if [[replicas]] is empty, read/write splitting will NOT be enabled no matter what value this property takes.
+     * @since 2.0.36
      */
-    public $enableSlaves = true;
+    public $enableReplicas = true;
     /**
-     * @var array list of slave connection configurations. Each configuration is used to create a slave DB connection.
-     * When [[enableSlaves]] is true, one of these configurations will be chosen and used to create a DB connection
+     * Returns the value of [[enableReplicas]].
+     * @return bool
+     * @deprecated since 2.0.36. Use [[enableReplicas]] instead.
+     */
+    public function getEnableSlaves()
+    {
+        return $this->enableReplicas;
+    }
+    /**
+     * Sets the value of [[enableReplicas]].
+     * @param bool $value
+     * @deprecated since 2.0.36. Use [[enableReplicas]] instead.
+     */
+    public function setEnableSlaves($value)
+    {
+        $this->enableReplicas = $value;
+    }
+    /**
+     * @var array list of replica connection configurations. Each configuration is used to create a replica DB connection.
+     * When [[enableReplicas]] is true, one of these configurations will be chosen and used to create a DB connection
      * for performing read queries only.
-     * @see enableSlaves
-     * @see slaveConfig
+     * @see enableReplicas
+     * @see replicaConfig
+     * @since 2.0.36
      */
-    public $slaves = [];
+    public $replicas = [];
     /**
-     * @var array the configuration that should be merged with every slave configuration listed in [[slaves]].
+     * Returns the value of [[replicas]].
+     * @return array
+     * @deprecated since 2.0.36. Use [[replicas]] instead.
+     */
+    public function getSlaves()
+    {
+        return $this->replicas;
+    }
+    /**
+     * Sets the value of [[replicas]].
+     * @param array $value
+     * @deprecated since 2.0.36. Use [[replicas]] instead.
+     */
+    public function setSlaves($value)
+    {
+        $this->replicas = $value;
+    }
+    /**
+     * @var array the configuration that should be merged with every replica configuration listed in [[replicas]].
      * For example,
      *
      * ```php
      * [
-     *     'username' => 'slave',
-     *     'password' => 'slave',
+     *     'username' => 'replica',
+     *     'password' => 'replica',
      *     'attributes' => [
      *         // use a smaller connection timeout
      *         PDO::ATTR_TIMEOUT => 10,
      *     ],
      * ]
      * ```
+     *
+     * @since 2.0.36
      */
-    public $slaveConfig = [];
+    public $replicaConfig = [];
     /**
-     * @var array list of master connection configurations. Each configuration is used to create a master DB connection.
+     * Returns the value of [[replicaConfig]].
+     * @return array
+     * @deprecated since 2.0.36. Use [[replicaConfig]] instead.
+     */
+    public function getSlaveConfig()
+    {
+        return $this->replicaConfig;
+    }
+    /**
+     * Sets the value of [[replicaConfig]].
+     * @param array $value
+     * @deprecated since 2.0.36. Use [[replicaConfig]] instead.
+     */
+    public function setSlaveConfig($value)
+    {
+        $this->replicaConfig = $value;
+    }
+    /**
+     * @var array list of primary connection configurations. Each configuration is used to create a primary DB connection.
      * When [[open()]] is called, one of these configurations will be chosen and used to create a DB connection
      * which will be used by this object.
-     * Note that when this property is not empty, the connection setting (e.g. "dsn", "username") of this object will
+     * Note that when this property is not empty, the connection setting (e.g. `dsn`, `username`) of this object will
      * be ignored.
-     * @see masterConfig
-     * @see shuffleMasters
+     * @see primaryConfig
+     * @see shufflePrimaries
+     * @since 2.0.36
      */
-    public $masters = [];
+    public $primaries = [];
     /**
-     * @var array the configuration that should be merged with every master configuration listed in [[masters]].
+     * Returns the value of [[primaries]].
+     * @return array
+     * @deprecated since 2.0.36. Use [[primaries]] instead.
+     */
+    public function getMasters()
+    {
+        return $this->primaries;
+    }
+    /**
+     * Sets the value of [[primaries]].
+     * @param array $value
+     * @deprecated since 2.0.36. Use [[primaries]] instead.
+     */
+    public function setMasters($value)
+    {
+        $this->primaries = $value;
+    }
+    /**
+     * @var array the configuration that should be merged with every primary configuration listed in [[primaries]].
      * For example,
      *
      * ```php
      * [
-     *     'username' => 'master',
-     *     'password' => 'master',
+     *     'username' => 'primary',
+     *     'password' => 'primary',
      *     'attributes' => [
      *         // use a smaller connection timeout
      *         PDO::ATTR_TIMEOUT => 10,
      *     ],
      * ]
      * ```
+     *
+     * @since 2.0.36
      */
-    public $masterConfig = [];
+    public $primaryConfig = [];
     /**
-     * @var bool whether to shuffle [[masters]] before getting one.
-     * @since 2.0.11
-     * @see masters
+     * Returns the value of [[primaryConfig]].
+     * @return array
+     * @deprecated since 2.0.36. Use [[primaryConfig]] instead.
      */
-    public $shuffleMasters = true;
+    public function getMasterConfig()
+    {
+        return $this->primaryConfig;
+    }
+    /**
+     * Sets the value of [[primaryConfig]].
+     * @param array $value
+     * @deprecated since 2.0.36. Use [[primaryConfig]] instead.
+     */
+    public function setMasterConfig($value)
+    {
+        $this->primaryConfig = $value;
+    }
+    /**
+     * @var bool whether to shuffle [[primaries]] before getting one.
+     * @since 2.0.11
+     * @see primaries
+     * @since 2.0.36
+     */
+    public $shufflePrimaries = true;
+    /**
+     * Returns the value of [[shufflePrimaries]].
+     * @return bool
+     * @deprecated since 2.0.36. Use [[shufflePrimaries]] instead.
+     */
+    public function getShuffleMasters()
+    {
+        return $this->shufflePrimaries;
+    }
+    /**
+     * Sets the value of [[shufflePrimaries]].
+     * @param bool $value
+     * @deprecated since 2.0.36. Use [[shufflePrimaries]] instead.
+     */
+    public function setShuffleMasters($value)
+    {
+        $this->shufflePrimaries = $value;
+    }
     /**
      * @var bool whether to enable logging of database queries. Defaults to true.
      * You may want to disable this option in a production environment to gain performance
@@ -432,13 +548,13 @@ class Connection extends Component
      */
     private $_driverName;
     /**
-     * @var Connection|false the currently active master connection
+     * @var Connection|false the currently active primary connection
      */
-    private $_master = false;
+    private $_primary = false;
     /**
-     * @var Connection|false the currently active slave connection
+     * @var Connection|false the currently active replica connection
      */
-    private $_slave = false;
+    private $_replica = false;
     /**
      * @var array query cache parameters for the [[cache()]] calls
      */
@@ -598,14 +714,14 @@ class Connection extends Component
             return;
         }
 
-        if (!empty($this->masters)) {
-            $db = $this->getMaster();
+        if (!empty($this->primaries)) {
+            $db = $this->getPrimary();
             if ($db !== null) {
                 $this->pdo = $db->pdo;
                 return;
             }
 
-            throw new InvalidConfigException('None of the master DB servers is available.');
+            throw new InvalidConfigException('None of the primary DB servers are available.');
         }
 
         if (empty($this->dsn)) {
@@ -644,13 +760,13 @@ class Connection extends Component
      */
     public function close()
     {
-        if ($this->_master) {
-            if ($this->pdo === $this->_master->pdo) {
+        if ($this->_primary) {
+            if ($this->pdo === $this->_primary->pdo) {
                 $this->pdo = null;
             }
 
-            $this->_master->close();
-            $this->_master = false;
+            $this->_primary->close();
+            $this->_primary = false;
         }
 
         if ($this->pdo !== null) {
@@ -658,9 +774,9 @@ class Connection extends Component
             $this->pdo = null;
         }
 
-        if ($this->_slave) {
-            $this->_slave->close();
-            $this->_slave = false;
+        if ($this->_replica) {
+            $this->_replica->close();
+            $this->_replica = false;
         }
 
         $this->_schema = null;
@@ -971,7 +1087,7 @@ class Connection extends Component
             if (($pos = strpos($this->dsn, ':')) !== false) {
                 $this->_driverName = strtolower(substr($this->dsn, 0, $pos));
             } else {
-                $this->_driverName = strtolower($this->getSlavePdo()->getAttribute(PDO::ATTR_DRIVER_NAME));
+                $this->_driverName = strtolower($this->getReplicaPdo()->getAttribute(PDO::ATTR_DRIVER_NAME));
             }
         }
 
@@ -998,79 +1114,139 @@ class Connection extends Component
     }
 
     /**
-     * Returns the PDO instance for the currently active slave connection.
-     * When [[enableSlaves]] is true, one of the slaves will be used for read queries, and its PDO instance
+     * Returns the PDO instance for the currently active replica connection.
+     * When [[enableReplicas]] is true, one of the replicas will be used for read queries, and its PDO instance
      * will be returned by this method.
-     * @param bool $fallbackToMaster whether to return a master PDO in case none of the slave connections is available.
-     * @return PDO the PDO instance for the currently active slave connection. `null` is returned if no slave connection
-     * is available and `$fallbackToMaster` is false.
+     * @param bool $fallbackToPrimary whether to return the primary PDO if no replica connections are available.
+     * @return PDO|null the PDO instance for the currently active replica connection. `null` is returned if no
+     * replica connections are available and `$fallbackToPrimary` is false.
+     * @since 2.0.36
      */
-    public function getSlavePdo($fallbackToMaster = true)
+    public function getReplicaPdo($fallbackToPrimary = true)
     {
-        $db = $this->getSlave(false);
+        $db = $this->getReplica(false);
         if ($db === null) {
-            return $fallbackToMaster ? $this->getMasterPdo() : null;
+            return $fallbackToPrimary ? $this->getPrimaryPdo() : null;
         }
 
         return $db->pdo;
     }
 
     /**
-     * Returns the PDO instance for the currently active master connection.
-     * This method will open the master DB connection and then return [[pdo]].
-     * @return PDO the PDO instance for the currently active master connection.
+     * Returns the PDO instance for the currently active replica connection.
+     * When [[enableReplicas]] is true, one of the replicas will be used for read queries, and its PDO instance
+     * will be returned by this method.
+     * @param bool $fallbackToPrimary whether to return the primary PDO if no replica connections are available.
+     * @return PDO|null the PDO instance for the currently active replica connection. `null` is returned if no
+     * replica connections are available and `$fallbackToPrimary` is false.
+     * @deprecated since 2.0.36. Use [[getReplicaPdo()]] instead.
      */
-    public function getMasterPdo()
+    public function getSlavePdo($fallbackToPrimary = true)
+    {
+        return $this->getReplicaPdo($fallbackToPrimary);
+    }
+
+    /**
+     * Returns the PDO instance for the currently active primary connection.
+     * This method will open the primary DB connection and then return [[pdo]].
+     * @return PDO the PDO instance for the currently active primary connection.
+     * @since 2.0.36
+     */
+    public function getPrimaryPdo()
     {
         $this->open();
         return $this->pdo;
     }
 
     /**
-     * Returns the currently active slave connection.
-     * If this method is called for the first time, it will try to open a slave connection when [[enableSlaves]] is true.
-     * @param bool $fallbackToMaster whether to return a master connection in case there is no slave connection available.
-     * @return Connection the currently active slave connection. `null` is returned if there is no slave available and
-     * `$fallbackToMaster` is false.
+     * Returns the PDO instance for the currently active primary connection.
+     * This method will open the primary DB connection and then return [[pdo]].
+     * @return PDO the PDO instance for the currently active primary connection.
+     * @deprecated since 2.0.36. Use [[getPrimaryPdo()]] instead.
      */
-    public function getSlave($fallbackToMaster = true)
+    public function getMasterPdo()
     {
-        if (!$this->enableSlaves) {
-            return $fallbackToMaster ? $this : null;
-        }
-
-        if ($this->_slave === false) {
-            $this->_slave = $this->openFromPool($this->slaves, $this->slaveConfig);
-        }
-
-        return $this->_slave === null && $fallbackToMaster ? $this : $this->_slave;
+        return $this->getPrimaryPdo();
     }
 
     /**
-     * Returns the currently active master connection.
-     * If this method is called for the first time, it will try to open a master connection.
-     * @return Connection the currently active master connection. `null` is returned if there is no master available.
+     * Returns the currently active replica connection.
+     * If this method is called for the first time, it will try to open a replica connection when [[enableReplicas]]
+     * is true.
+     * @param bool $fallbackToPrimary whether to return the primary connection if no replica connections are
+     * available.
+     * @return Connection|null the currently active replica connection. `null` is returned if no replica connections
+     * are available and `$fallbackToPrimary` is false.
+     * @since 2.0.36
+     */
+    public function getReplica($fallbackToPrimary = true)
+    {
+        if (!$this->enableReplicas) {
+            return $fallbackToPrimary ? $this : null;
+        }
+
+        if ($this->_replica === false) {
+            $this->_replica = $this->openFromPool($this->replicas, $this->replicaConfig);
+        }
+
+        return $this->_replica === null && $fallbackToPrimary ? $this : $this->_replica;
+    }
+
+    /**
+     * Returns the currently active replica connection.
+     * If this method is called for the first time, it will try to open a replica connection when [[enableReplicas]]
+     * is true.
+     * @param bool $fallbackToPrimary whether to return the primary connection if no replica connections are
+     * available.
+     * @return Connection|null the currently active replica connection. `null` is returned if no replica connections
+     * are available and `$fallbackToPrimary` is false.
+     * @deprecated since 2.0.36. Use [[getReplica()]] instead.
+     */
+    public function getSlave($fallbackToPrimary = true)
+    {
+        return $this->getReplica($fallbackToPrimary);
+    }
+
+
+    /**
+     * Returns the currently active primary connection.
+     * If this method is called for the first time, it will try to open a primary connection.
+     * @return Connection|null the currently active primary connection. `null` is returned if no primary connection
+     * is available.
+     * @since 2.0.36
+     */
+    public function getPrimary()
+    {
+        if ($this->_primary === false) {
+            $this->_primary = $this->shufflePrimaries
+                ? $this->openFromPool($this->primaries, $this->primaryConfig)
+                : $this->openFromPoolSequentially($this->primaries, $this->primaryConfig);
+        }
+
+        return $this->_primary;
+    }
+
+    /**
+     * Returns the currently active primary connection.
+     * If this method is called for the first time, it will try to open a primary connection.
+     * @return Connection|null the currently active primary connection. `null` is returned if no primary connection
+     * is available.
      * @since 2.0.11
+     * @deprecated since 2.0.36. Use [[getPrimary()]] instead.
      */
     public function getMaster()
     {
-        if ($this->_master === false) {
-            $this->_master = $this->shuffleMasters
-                ? $this->openFromPool($this->masters, $this->masterConfig)
-                : $this->openFromPoolSequentially($this->masters, $this->masterConfig);
-        }
-
-        return $this->_master;
+        return $this->getPrimary();
     }
 
     /**
-     * Executes the provided callback by using the master connection.
+     * Executes the provided callback by using the primary connection.
      *
-     * This method is provided so that you can temporarily force using the master connection to perform
+     * This method is provided so that you can temporarily force using the primary connection to perform
      * DB operations even if they are read queries. For example,
      *
      * ```php
-     * $result = $db->useMaster(function ($db) {
+     * $result = $db->usePrimary(function ($db) {
      *     return $db->createCommand('SELECT * FROM user LIMIT 1')->queryOne();
      * });
      * ```
@@ -1079,27 +1255,51 @@ class Connection extends Component
      * `function (Connection $db)`. Its return value will be returned by this method.
      * @return mixed the return value of the callback
      * @throws \Exception|\Throwable if there is any exception thrown from the callback
+     * @since 2.0.36
      */
-    public function useMaster(callable $callback)
+    public function usePrimary(callable $callback)
     {
-        if ($this->enableSlaves) {
-            $this->enableSlaves = false;
+        if ($this->enableReplicas) {
+            $this->enableReplicas = false;
             try {
                 $result = call_user_func($callback, $this);
             } catch (\Exception $e) {
-                $this->enableSlaves = true;
+                $this->enableReplicas = true;
                 throw $e;
             } catch (\Throwable $e) {
-                $this->enableSlaves = true;
+                $this->enableReplicas = true;
                 throw $e;
             }
             // TODO: use "finally" keyword when miminum required PHP version is >= 5.5
-            $this->enableSlaves = true;
+            $this->enableReplicas = true;
         } else {
             $result = call_user_func($callback, $this);
         }
 
         return $result;
+    }
+
+    /**
+     * Executes the provided callback by using the primary connection.
+     *
+     * This method is provided so that you can temporarily force using the primary connection to perform
+     * DB operations even if they are read queries. For example,
+     *
+     * ```php
+     * $result = $db->usePrimary(function ($db) {
+     *     return $db->createCommand('SELECT * FROM user LIMIT 1')->queryOne();
+     * });
+     * ```
+     *
+     * @param callable $callback a PHP callable to be executed by this method. Its signature is
+     * `function (Connection $db)`. Its return value will be returned by this method.
+     * @return mixed the return value of the callback
+     * @throws \Exception|\Throwable if there is any exception thrown from the callback
+     * @deprecated since 2.0.36. Use [[usePrimary()]] instead.
+     */
+    public function useMaster(callable $callback)
+    {
+        return $this->usePrimary($callback);
     }
 
     /**
@@ -1219,8 +1419,8 @@ class Connection extends Component
         $fields = (array) $this;
 
         unset($fields['pdo']);
-        unset($fields["\000" . __CLASS__ . "\000" . '_master']);
-        unset($fields["\000" . __CLASS__ . "\000" . '_slave']);
+        unset($fields["\000" . __CLASS__ . "\000" . '_primary']);
+        unset($fields["\000" . __CLASS__ . "\000" . '_replica']);
         unset($fields["\000" . __CLASS__ . "\000" . '_transaction']);
         unset($fields["\000" . __CLASS__ . "\000" . '_schema']);
 
@@ -1234,8 +1434,8 @@ class Connection extends Component
     {
         parent::__clone();
 
-        $this->_master = false;
-        $this->_slave = false;
+        $this->_primary = false;
+        $this->_replica = false;
         $this->_schema = null;
         $this->_transaction = null;
         if (strncmp($this->dsn, 'sqlite::memory:', 15) !== 0) {