+++ framework/db/oci/QueryBuilder.php
@@ -154,8 +154,8 @@ EOD;
             if (count($tableSchema->primaryKey)>1) {
                 throw new InvalidArgumentException("Can't reset sequence for composite primary key in table: $table");
             }
-            // use master connection to get the biggest PK value
-            $value = $this->db->useMaster(function (Connection $db) use ($tableSchema) {
+            // use primary connection to get the biggest PK value
+            $value = $this->db->usePrimary(function (Connection $db) use ($tableSchema) {
                 return $db->createCommand(
                     'SELECT MAX("' . $tableSchema->primaryKey[0] . '") FROM "'. $tableSchema->name . '"'
                 )->queryScalar();