+++ framework/CHANGELOG.md
@@ -15,6 +15,20 @@ Yii Framework 2 Change Log
 - Bug #13828: Fix retrieving inserted data for a primary key of type uniqueidentifier for SQL Server 2005 or later (darkdef)
 - Bug #17474: Fix retrieving inserted data for a primary key of type trigger for SQL Server 2005 or later (darkdef)
 - Bug #18001: Fix getting table metadata for tables `(` in their name (floor12)
+- Enh #18102: Use “primary”/“replica” terminology instead of “master”/“slave” (brandonkelly)
+  - Added `yii\db\Connection::$enableReplicas` and deprecated `$enableSlaves` via magic methods.
+  - Added `yii\db\Connection::$replicas` and deprecated `$slaves` via magic methods.
+  - Added `yii\db\Connection::$replicaConfig` and deprecated `$slaveConfig` via magic methods.
+  - Added `yii\db\Connection::$primaries` and deprecated `$masters` via magic methods.
+  - Added `yii\db\Connection::$primaryConfig` and deprecated `$masterConfig` via magic methods.
+  - Added `yii\db\Connection::$shufflePrimaries` and deprecated `$shuffleMasters` via magic methods.
+  - Added `yii\db\Connection::getReplicaPdo()` and deprecated `getSlavePdo()`.
+  - Added `yii\db\Connection::getPrimaryPdo()` and deprecated `getMasterPdo()`.
+  - Added `yii\db\Connection::getReplica()` and deprecated `getSlave()`.
+  - Added `yii\db\Connection::getPrimary()` and deprecated `getMaster()`.
+  - Added `yii\db\Connection::usePrimary()` and deprecated `useMaster()`.
+  - Added `yii\validators\ExistValidator::$forcePrimaryDb` and deprecated `$forceMasterDb` via magic methods.
+  - Added `yii\validators\UniqueValidator::$forcePrimaryDb` and deprecated `$forceMasterDb` via magic methods.
 
 
 2.0.35 May 02, 2020